pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nltouml"
version = "0.1.2"
description = "Natural Language to PlantUML State Machine (IR-centered) pipeline"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "python-dotenv>=1.0.0",
  "jsonschema>=4.20.0",
  "pydantic>=2.0.0",
]

[project.optional-dependencies]
openai = ["openai>=1.0.0"]

[project.scripts]
nltouml = "nltouml.cli:main"
nlpipeline = "nltouml.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
nltouml = ["templates/*.json"]
